+++ src/main/java/haja/Project/api/TasknoticeApiController.java
@@ -10,6 +10,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.annotation.CreatedDate;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.io.File;
@@ -35,43 +37,48 @@ public class TasknoticeApiController {
     @Operation(summary = "과제 공지사항 생성")
     @PostMapping("tasknotice")
     public CreateTasknoticeResponse createTasknotice(@RequestBody @Valid CreateTasknoticeRequest request){
-        Tasknotice tasknotice = new Tasknotice();
-        //tag_id만 리스트로 받고
-        //태그버튼을 누르면 tasknotice_tag객체가 생성되게 하고
-        tasknotice.setMember(memberService.findById(SecurityUtil.getCurrentMemberId()).get());
-        //System.out.println(tasknotice.getMember().getEmail()); -> setMember 잘 됐는지 테스트용/ 잘 되는거 확인했습니다
-        tasknotice.setDate(LocalDateTime.now()); //생성시점
-        tasknotice.setDeadline(request.getDeadline()); // postman으로 날짜받는거
-        tasknotice.setTarget(request.getTarget());
-        tasknotice.setTitle(request.getTitle());
-        tasknotice.setExplanation(request.getExplanation());
+        if(memberService.findById(SecurityUtil.getCurrentMemberId()).get().getAuthority() == Authority.ROLE_ADMIN) {
+        //System.out.println("->" + memberService.findById(SecurityUtil.getCurrentMemberId()).get().getAuthority());
+        //-> 위의 결과 ROLE_USER로 나옴
+            Tasknotice tasknotice = new Tasknotice();
+            //tag_id만 리스트로 받고
+            //태그버튼을 누르면 tasknotice_tag객체가 생성되게 하고
+            tasknotice.setMember(memberService.findById(SecurityUtil.getCurrentMemberId()).get());
+            //System.out.println(tasknotice.getMember().getEmail()); -> setMember 잘 됐는지 테스트용/ 잘 되는거 확인했습니다
+            tasknotice.setDate(LocalDateTime.now()); //생성시점
+            tasknotice.setDeadline(request.getDeadline()); // postman으로 날짜받는거
+            tasknotice.setTarget(request.getTarget());
+            tasknotice.setTitle(request.getTitle());
+            tasknotice.setExplanation(request.getExplanation());
 
-        Long id = tasknoticeService.save(tasknotice);
+            Long id = tasknoticeService.save(tasknotice);
 
-        // 태그 저장
-        List<String> tags = request.tags;
-        if (tags != null) {
-            for (String tag_name : tags) {
-                if (tagService.findByName(tag_name) == null) {  //태그가 없는 태그면 새로생성해서
-                    Tag tag = new Tag();
-                    tag.setName(tag_name);
-                    tagService.save(tag);
+            // 태그 저장
+            List<String> tags = request.tags;
+            if (tags != null) {
+                for (String tag_name : tags) {
+                    if (tagService.findByName(tag_name) == null) {  //태그가 없는 태그면 새로생성해서
+                        Tag tag = new Tag();
+                        tag.setName(tag_name);
+                        tagService.save(tag);
 
-                    Tasknotice_Tag tasknotice_tag = new Tasknotice_Tag();
-                    tasknotice_tag.setTasknotice(tasknoticeService.findOne(id));
-                    tasknotice_tag.setTag(tag);
-                    tasknotice_tagService.save(tasknotice_tag);
-                }
-                else {   //이미 있는 태그면 걍 바로 넣어줌
-                    Tasknotice_Tag tasknotice_tag = new Tasknotice_Tag();
-                    tasknotice_tag.setTasknotice(tasknoticeService.findOne(id));
-                    tasknotice_tag.setTag(tagService.findByName(tag_name));
-                    tasknotice_tagService.save(tasknotice_tag);
+                        Tasknotice_Tag tasknotice_tag = new Tasknotice_Tag();
+                        tasknotice_tag.setTasknotice(tasknoticeService.findOne(id));
+                        tasknotice_tag.setTag(tag);
+                        tasknotice_tagService.save(tasknotice_tag);
+                    } else {   //이미 있는 태그면 걍 바로 넣어줌
+                        Tasknotice_Tag tasknotice_tag = new Tasknotice_Tag();
+                        tasknotice_tag.setTasknotice(tasknoticeService.findOne(id));
+                        tasknotice_tag.setTag(tagService.findByName(tag_name));
+                        tasknotice_tagService.save(tasknotice_tag);
+                    }
                 }
             }
-        }
 
-        return new CreateTasknoticeResponse(id);  // new 조심
+            return new CreateTasknoticeResponse(id);  // new 조심
+        }
+        return null;
+
     }
     @Data
     static class CreateTasknoticeRequest{
@@ -104,20 +111,23 @@ public class TasknoticeApiController {
     @GetMapping("tasknotice/{id}")
     public TasknoticeDto tasknotice(
             @PathVariable("id") Long id){
+        if(memberService.findById(SecurityUtil.getCurrentMemberId()).get().getAuthority() == Authority.ROLE_ADMIN) {
+            TasknoticeDto d = new TasknoticeDto(tasknoticeService.findOne(id)); //이렇게 하면 Json이 그대로 오는것을 확인함
+            //그니까 수정하기 버튼 누르면 이렇게 주고
+            //수정완료 버튼 누르면 request날라온거 set으로 ㄱㄱ
+            if (taskService.isSubmit(d.id)) d.isSubmit = true;
+            else d.isSubmit = false;
 
-        TasknoticeDto d = new TasknoticeDto(tasknoticeService.findOne(id)); //이렇게 하면 Json이 그대로 오는것을 확인함
-        //그니까 수정하기 버튼 누르면 이렇게 주고
-        //수정완료 버튼 누르면 request날라온거 set으로 ㄱㄱ
-        if (taskService.isSubmit(d.id)) d.isSubmit = true;
-        else d.isSubmit = false;
+            // 태그 추가
+            List<Tasknotice_Tag> tags = tasknotice_tagService.findByTasknoticeId(d.id);
+            for (Tasknotice_Tag tag : tags) {
+                d.tag.add(tag.getTag().getName());
+            }
 
-        // 태그 추가
-        List<Tasknotice_Tag> tags = tasknotice_tagService.findByTasknoticeId(d.id);
-        for (Tasknotice_Tag tag: tags) {
-            d.tag.add(tag.getTag().getName());
+            return d;
         }
-
-        return d;
+        else
+            return null;
 
     }
 
@@ -127,51 +137,54 @@ public class TasknoticeApiController {
     public UpdateResponse tasknotice(
             @PathVariable("id") Long id,
             @RequestBody @Valid UpdateRequest request){
-        Tasknotice tn;
-        tn = tasknoticeService.findOne(id);
+        if(memberService.findById(SecurityUtil.getCurrentMemberId()).get().getAuthority() == Authority.ROLE_ADMIN) {
+            Tasknotice tn;
+            tn = tasknoticeService.findOne(id);
 
-        //updateTime 도 하나 만들어야할듯
-        tn.setDeadline(request.getDeadline()); // postman으로 날짜받는거
-        tn.setTarget(request.getTarget());
-        tn.setTitle(request.getTitle());
-        tn.setExplanation(request.getExplanation());
+            //updateTime 도 하나 만들어야할듯
+            tn.setDeadline(request.getDeadline()); // postman으로 날짜받는거
+            tn.setTarget(request.getTarget());
+            tn.setTitle(request.getTitle());
+            tn.setExplanation(request.getExplanation());
 
-        Long tn_id = tasknoticeService.save(tn);
-        //이러면 이제 과제공지글을 생성이 된거고 이 때 request로 tag_id받고 아래에 tasknoticetag로직 추가
-        //request로 tag_id를 어떻게받을까 List로 받아서 for문으로 돌리자
+            Long tn_id = tasknoticeService.save(tn);
+            //이러면 이제 과제공지글을 생성이 된거고 이 때 request로 tag_id받고 아래에 tasknoticetag로직 추가
+            //request로 tag_id를 어떻게받을까 List로 받아서 for문으로 돌리자
 
-        List<Tasknotice_Tag> tts = tasknotice_tagService.findByTasknoticeId(id); //여기부터 tasknotice_tag삭제 후 생성
-        List<ttDTO> result = tts.stream()
-                .map(t -> new ttDTO(t))
-                .collect(Collectors.toList());
+            List<Tasknotice_Tag> tts = tasknotice_tagService.findByTasknoticeId(id); //여기부터 tasknotice_tag삭제 후 생성
+            List<ttDTO> result = tts.stream()
+                    .map(t -> new ttDTO(t))
+                    .collect(Collectors.toList());
 
-        for(int i=0; i<result.size();i++) {
-            Tasknotice_Tag tt = tasknotice_tagService.findOne(result.get(i).id);
-            tasknotice_tagService.delete(tt);
-        }
+            for (int i = 0; i < result.size(); i++) {
+                Tasknotice_Tag tt = tasknotice_tagService.findOne(result.get(i).id);
+                tasknotice_tagService.delete(tt);
+            }
 
-        List<String> tags = request.tags;
-        if (tags != null) {
-            for (String tag_name : tags) {
-                if (tagService.findByName(tag_name) == null) {
-                    Tag tag = new Tag();
-                    tag.setName(tag_name);
-                    tagService.save(tag);
+            List<String> tags = request.tags;
+            if (tags != null) {
+                for (String tag_name : tags) {
+                    if (tagService.findByName(tag_name) == null) {
+                        Tag tag = new Tag();
+                        tag.setName(tag_name);
+                        tagService.save(tag);
 
-                    Tasknotice_Tag tasknotice_tag = new Tasknotice_Tag();
-                    tasknotice_tag.setTasknotice(tasknoticeService.findOne(tn_id));
-                    tasknotice_tag.setTag(tag);
-                    tasknotice_tagService.save(tasknotice_tag);
-                }
-                else {
-                    Tasknotice_Tag tasknotice_tag = new Tasknotice_Tag();
-                    tasknotice_tag.setTasknotice(tasknoticeService.findOne(tn_id));
-                    tasknotice_tag.setTag(tagService.findByName(tag_name));
-                    tasknotice_tagService.save(tasknotice_tag);
+                        Tasknotice_Tag tasknotice_tag = new Tasknotice_Tag();
+                        tasknotice_tag.setTasknotice(tasknoticeService.findOne(tn_id));
+                        tasknotice_tag.setTag(tag);
+                        tasknotice_tagService.save(tasknotice_tag);
+                    } else {
+                        Tasknotice_Tag tasknotice_tag = new Tasknotice_Tag();
+                        tasknotice_tag.setTasknotice(tasknoticeService.findOne(tn_id));
+                        tasknotice_tag.setTag(tagService.findByName(tag_name));
+                        tasknotice_tagService.save(tasknotice_tag);
+                    }
                 }
             }
+            return new UpdateResponse(tn_id);
         }
-        return new UpdateResponse(tn_id);
+        else
+            return null;
     }
 
     @Data
@@ -205,8 +218,10 @@ public class TasknoticeApiController {
     @Operation(summary = "과제 공지사항 삭제")
     @DeleteMapping("tasknotice/{id}")
     public void deleteTasknotice(@PathVariable("id") Long id) {
-        tasknotice_tagService.deleteByTasknoticeId(id);
-        tasknoticeService.delete(id);
+        if(memberService.findById(SecurityUtil.getCurrentMemberId()).get().getAuthority() == Authority.ROLE_ADMIN) {
+            tasknotice_tagService.deleteByTasknoticeId(id);
+            tasknoticeService.delete(id);
+        }
     }
